fixed a bug how changed file is deleted
Previously file was never actually removed since file.storage
was always empty due to a fact how Django pickles FileField.
Normally FileFieldDescriptor would populate the necessary
attributes back however since in order to detect which fields
change during save, they need to be copied which means that
storage attribute is lost and so the file was not actually
removed. In this fix we explicitly copy the necessary parameters
to the old file field so that it can successfully be removed
by the appropriate storage mechanism.